Add inventory plugin to list active VMs
This adds a new task `floaty::list` which is intended for use as a Bolt
inventory plugin. It lists active VMs acquired through floaty and
returns the list of bare hostnames under the 'value' key. It does this
by running `floaty list --active` in a new process and munging the
output to a list of hostnames. The task doesn't currently accept
parameters, requiring users to configure their connection to vmpooler in
`~/.vmfloaty.yml`. This should be safe for now, as specifying connection
information on the command line is painful and it's expected most users
already do this.

This task allows floaty users to run bolt against their VMs
without making daily changes to their inventory file or manually
specifying the VM hostnames.