(maint) Handle using abs or vmpooler service
The `list` task now accepts an optional `service` parameter to specify
which pooler service to list active VMs from. If the parameter isn't
provided the task tries to read `~/.vmfloaty.yml` if it exists to
determine which service floaty will use. If it can't determine the
service from the configfile it falls back to `vmpooler`. The task can
then handle the output format for the ABS service and VMPooler service,
which have different structures of output.